requirements.txt
fastmcp
oci
oracle-ads
requests
pydantic
uvicorn
starlette
pandas